Add native format function & pass args
I've no idea how to make Native modules and the guide I followed might
be out of date but this is a start.

The idea is that we use a native module in order to directly expose the
messageformat node module that I do not wish to re-implement in Elm.

We improve the dumped elm code so that it includes arguments in the
types and calls the MessageFormat function.