Update code generation for languages
We want to be able to have language strings in different files and be
aware of the corresponding language code for them. We previously had the
language codes hard coded in the generation but now we base it off a
config file so that we can easily manage the details and possibly extent
with other settings.

The generation code isn't that pretty. Should clean it up a bit at some
point.