Add diff action
So that we can diff the expected output of config file against a current
file to find out if something needs to change.

Designed to be used in tests to make sure that your elm code is in sync
with your config & language files. Though really this should be exposed
as some kind of module with api rather than a command line call. Early
days though.