Add basic filtering support
This is for when we want to only extract a subset of the json strings
for processing to elm code. There is a solid change that the language
file will include more strings than are required for the elm app so we
try to provide an option to filter down to only what is needed.

We also use 'commander' for options parsing and 'lodash' so that we
don't need a bleeding edge node version with Object.entries and also to
make the filtering code easier.