Allow for 'context' in translation file
This is a little direct and should be refactored and made more general.
The idea is that previously the language file was a straight key to
string json object, but I'd like the language file to have 'context'
along side the value so the language file becomes a map from 'keys' to
little objects with 'value' & 'context'. For the purposes of this script
we just need the 'value' and can ignore the context and carry on.