Add language lookup object
We want a way to abstract away from using the raw Translation tokens as
those are part of a type and that means that if a shared module wants to
use translation tokens it has to have its own token type or all apps
that use the module have to have the same token type.

Introducing this allows us to have an indirection between the references
and the tokens. Shared code can refer to properties on an object using
.accessor functions and not worry about the rest of the object and then
apps using this shared code only have to worry about providing an object
has at least the required entries rather than having a strictly
identicial type.